chore: AddressMap should be generic to MemoryBackend
