+++ crates/vm/src/system/memory/online/memmap.rs
@@ -1,16 +1,6 @@
-use std::{
-    fmt::Debug,
-    marker::PhantomData,
-    mem::MaybeUninit,
-    ptr::{copy_nonoverlapping, slice_from_raw_parts},
-};
+use std::{fmt::Debug, marker::PhantomData, mem::MaybeUninit, ptr::copy_nonoverlapping};
 
-use itertools::{zip_eq, Itertools};
 use memmap2::MmapMut;
-use openvm_instructions::exe::SparseMemoryImage;
-use openvm_stark_backend::p3_field::PrimeField32;
-
-use super::{Address, MemoryConfig};
 
 pub const CELL_STRIDE: usize = 1;
 /// Default mmap page size. Change this if using THB.
@@ -191,168 +181,5 @@ impl MmapMemory {
         // - All elements have been initialized (zero-initialized if didn't exist).
         // - `result` is aligned to `BLOCK`
         unsafe { result.assume_init() }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct AddressMap {
-    pub mem: Vec<MmapMemory>,
-    pub cell_size: Vec<usize>, // TODO: move to MmapWrapper
-}
-
-impl Default for AddressMap {
-    fn default() -> Self {
-        Self::from_mem_config(&MemoryConfig::default())
-    }
-}
-
-impl AddressMap {
-    pub fn new(mem_size: Vec<usize>) -> Self {
-        // TMP: hardcoding for now
-        let mut cell_size = vec![1; 4];
-        cell_size.resize(mem_size.len(), 4);
-        let mem = zip_eq(&cell_size, &mem_size)
-            .map(|(cell_size, mem_size)| MmapMemory::new(mem_size.checked_mul(*cell_size).unwrap()))
-            .collect();
-        Self { mem, cell_size }
-    }
-
-    pub fn from_mem_config(mem_config: &MemoryConfig) -> Self {
-        Self::new(mem_config.addr_space_sizes.clone())
-    }
-
-    pub fn get_memory(&self) -> &Vec<MmapMemory> {
-        &self.mem
-    }
-
-    pub fn get_memory_mut(&mut self) -> &mut Vec<MmapMemory> {
-        &mut self.mem
-    }
-
-    pub fn items<F: PrimeField32>(&self) -> impl Iterator<Item = (Address, F)> + '_ {
-        zip_eq(&self.mem, &self.cell_size).enumerate().flat_map(
-            move |(as_idx, (space_mem, &cell_size))| {
-                // TODO: better way to handle address space conversions to F
-                if cell_size == 1 {
-                    space_mem
-                        .iter::<u8>()
-                        .map(move |(ptr_idx, x)| {
-                            ((as_idx as u32, ptr_idx as u32), F::from_canonical_u8(x))
-                        })
-                        .collect_vec()
-                } else {
-                    assert_eq!(cell_size, 4);
-                    space_mem
-                        .iter::<F>()
-                        .map(move |(ptr_idx, x)| ((as_idx as u32, ptr_idx as u32), x))
-                        .collect_vec()
-                }
-            },
-        )
-    }
-
-    pub fn get_f<F: PrimeField32>(&self, addr_space: u32, ptr: u32) -> F {
-        debug_assert_ne!(addr_space, 0);
-        // TODO: fix this
-        unsafe {
-            if self.cell_size[addr_space as usize] == 1 {
-                F::from_canonical_u8(self.get::<u8>((addr_space, ptr)))
-            } else {
-                debug_assert_eq!(self.cell_size[addr_space as usize], 4);
-                self.get::<F>((addr_space, ptr))
-            }
-        }
-    }
-
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    pub unsafe fn get<T: Copy>(&self, (addr_space, ptr): Address) -> T {
-        debug_assert_eq!(size_of::<T>(), self.cell_size[addr_space as usize]);
-        self.mem
-            .get_unchecked(addr_space as usize)
-            .get((ptr as usize) * size_of::<T>())
-    }
-
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    /// - ASSUMES `ptr..ptr + len` must not be out of bounds
-    pub unsafe fn get_slice<T: Copy + Debug>(
-        &self,
-        (addr_space, ptr): Address,
-        len: usize,
-    ) -> &[T] {
-        debug_assert_eq!(size_of::<T>(), self.cell_size[addr_space as usize]);
-        let start = (ptr as usize) * size_of::<T>();
-        let mem = self.mem.get_unchecked(addr_space as usize);
-        let slice_ptr = mem.as_ptr().add(start);
-        &*slice_from_raw_parts(slice_ptr as *const T, len)
-    }
-
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    pub unsafe fn insert<T: Copy>(&mut self, (addr_space, ptr): Address, data: T) -> T {
-        debug_assert_eq!(size_of::<T>(), self.cell_size[addr_space as usize]);
-        self.mem[addr_space as usize].replace(ptr as usize, &data)
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.mem.iter().all(|mem| mem.is_empty())
-    }
-
-    pub fn read_range_generic<T: Copy + Debug>(
-        &self,
-        (addr_space, ptr): Address,
-        len: usize,
-    ) -> Vec<T> {
-        let mut block: Vec<T> = Vec::with_capacity(len);
-        unsafe {
-            self.mem[addr_space as usize].read_range_generic(
-                ptr as usize,
-                len,
-                block.as_mut_ptr() as *mut u8,
-            );
-            block.set_len(len);
-        }
-        block
-    }
-
-    /// Copies `data` into the memory at `(addr_space, ptr)`.
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and `ptr + size_of_val(data)` is not
-    ///   out of bounds
-    pub unsafe fn copy_slice_nonoverlapping<T: Copy>(
-        &mut self,
-        (addr_space, ptr): Address,
-        data: &[T],
-    ) {
-        let start = (ptr as usize) * size_of::<T>();
-        let len = size_of_val(data);
-
-        // SAFETY:
-        // - `data` size is correct by definition
-        // - we assume `mem` will not go out of bounds
-        // - aligment is trivial for u8
-        // - `data` and `self.mem` are non-overlapping
-        std::ptr::copy_nonoverlapping(
-            data.as_ptr() as *const u8,
-            self.mem[addr_space as usize].as_mut_ptr().add(start),
-            len,
-        );
-    }
-
-    // TODO[jpw]: stabilize the boundary memory image format and how to construct
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    pub fn from_sparse(mem_size: Vec<usize>, sparse_map: SparseMemoryImage) -> Self {
-        let mut vec = Self::new(mem_size);
-        for ((addr_space, index), data_byte) in sparse_map.into_iter() {
-            vec.mem[addr_space as usize].set(index as usize, &data_byte);
-        }
-        vec
     }
 }
