+++ crates/vm/src/system/memory/online/paged_vec.rs
@@ -1,12 +1,6 @@
 use std::{fmt::Debug, marker::PhantomData, mem::MaybeUninit, ptr};
 
-use itertools::{zip_eq, Itertools};
-use openvm_instructions::exe::SparseMemoryImage;
-use openvm_stark_backend::p3_field::PrimeField32;
 use serde::{Deserialize, Serialize};
-
-use super::{Address, MemoryConfig};
-use crate::arch::ADDR_SPACE_OFFSET;
 
 /// 4096 is the default page size on host architectures if huge pages is not enabled
 const PAGE_SIZE: usize = 1 << 12;
@@ -284,346 +278,4 @@ impl<T: Copy> Iterator for PagedVecIter<'_, T> {
         }
         Some((global_index / size_of::<T>(), value))
     }
-}
-
-/// Map from address space to guest memory.
-/// The underlying memory is typeless, stored as raw bytes, but usage
-/// implicitly assumes that each address space has memory cells of a fixed type (e.g., `u8, F`).
-/// We do not use a typemap for performance reasons, and it is up to the user to enforce types.
-/// Needless to say, this is a very `unsafe` API.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AddressMap {
-    pub paged_vecs: Vec<PagedVec>,
-    /// byte size of cells per address space
-    pub cell_size: Vec<usize>,
-}
-
-impl Default for AddressMap {
-    fn default() -> Self {
-        Self::from_mem_config(&MemoryConfig::default())
-    }
-}
-
-impl AddressMap {
-    pub fn new(mem_size: Vec<usize>) -> Self {
-        // TMP: hardcoding for now
-        let mut cell_size = vec![1; 4];
-        cell_size.resize(mem_size.len(), 4);
-        let paged_vecs = zip_eq(&cell_size, &mem_size)
-            .map(|(cell_size, mem_size)| {
-                PagedVec::new(
-                    mem_size
-                        .checked_mul(*cell_size)
-                        .unwrap()
-                        .div_ceil(PAGE_SIZE),
-                )
-            })
-            .collect();
-        Self {
-            paged_vecs,
-            cell_size,
-        }
-    }
-    pub fn from_mem_config(mem_config: &MemoryConfig) -> Self {
-        Self::new(mem_config.addr_space_sizes.clone())
-    }
-
-    pub fn get_memory(&self) -> &Vec<PagedVec> {
-        &self.paged_vecs
-    }
-
-    pub fn get_memory_mut(&mut self) -> &mut Vec<PagedVec> {
-        &mut self.paged_vecs
-    }
-
-    pub fn items<F: PrimeField32>(&self) -> impl Iterator<Item = (Address, F)> + '_ {
-        zip_eq(&self.paged_vecs, &self.cell_size)
-            .enumerate()
-            .flat_map(move |(as_idx, (page, &cell_size))| {
-                // TODO: better way to handle address space conversions to F
-                if cell_size == 1 {
-                    page.iter::<u8>()
-                        .map(move |(ptr_idx, x)| {
-                            ((as_idx as u32, ptr_idx as u32), F::from_canonical_u8(x))
-                        })
-                        .collect_vec()
-                } else {
-                    // TEMP
-                    assert_eq!(cell_size, 4);
-                    page.iter::<F>()
-                        .map(move |(ptr_idx, x)| ((as_idx as u32, ptr_idx as u32), x))
-                        .collect_vec()
-                }
-            })
-    }
-
-    pub fn get_f<F: PrimeField32>(&self, addr_space: u32, ptr: u32) -> F {
-        debug_assert_ne!(addr_space, 0);
-        // TODO: fix this
-        unsafe {
-            if addr_space <= 3 {
-                F::from_canonical_u8(self.get::<u8>((addr_space, ptr)))
-            } else {
-                self.get::<F>((addr_space, ptr))
-            }
-        }
-    }
-
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    pub unsafe fn get<T: Copy>(&self, (addr_space, ptr): Address) -> T {
-        debug_assert_eq!(size_of::<T>(), self.cell_size[addr_space as usize]);
-        self.paged_vecs
-            .get_unchecked(addr_space as usize)
-            .get((ptr as usize) * size_of::<T>())
-    }
-
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    /// - Panics if accessing out of bounds memory within the address space
-    pub unsafe fn get_slice<T: Copy + Debug>(
-        &self,
-        (addr_space, ptr): Address,
-        len: usize,
-    ) -> Vec<T> {
-        let mut block: Vec<T> = Vec::with_capacity(len);
-        unsafe {
-            self.paged_vecs
-                .get_unchecked(addr_space as usize)
-                .read_range_generic(
-                    (ptr as usize) * size_of::<T>(),
-                    len * size_of::<T>(),
-                    block.as_mut_ptr() as *mut u8,
-                );
-            block.set_len(len);
-        }
-        block
-    }
-
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    pub unsafe fn insert<T: Copy>(&mut self, (addr_space, ptr): Address, data: T) -> T {
-        debug_assert_eq!(size_of::<T>(), self.cell_size[addr_space as usize]);
-        self.paged_vecs
-            .get_unchecked_mut(addr_space as usize)
-            .replace((ptr as usize) * size_of::<T>(), &data)
-    }
-    pub fn is_empty(&self) -> bool {
-        self.paged_vecs.iter().all(|page| page.is_empty())
-    }
-
-    /// Copies `data` into the memory at `(addr_space, ptr)`.
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    pub unsafe fn copy_slice_nonoverlapping<T: Copy>(
-        &mut self,
-        (addr_space, ptr): Address,
-        data: &[T],
-    ) {
-        let start = (ptr as usize) * size_of::<T>();
-        let len = size_of_val(data);
-
-        self.paged_vecs
-            .get_unchecked_mut((addr_space - ADDR_SPACE_OFFSET) as usize)
-            .set_range_generic(start, len, data.as_ptr() as *const u8);
-    }
-
-    // TODO[jpw]: stabilize the boundary memory image format and how to construct
-    /// # Safety
-    /// - `T` **must** be the correct type for a single memory cell for `addr_space`
-    /// - Assumes `addr_space` is within the configured memory and not out of bounds
-    pub fn from_sparse(mem_size: Vec<usize>, sparse_map: SparseMemoryImage) -> Self {
-        let mut vec = Self::new(mem_size);
-        for ((addr_space, index), data_byte) in sparse_map.into_iter() {
-            vec.paged_vecs[addr_space as usize].set(index as usize, &data_byte);
-        }
-        vec
-    }
-}
-
-// impl<const PAGE_SIZE: usize> GuestMemory for AddressMap<PAGE_SIZE> {
-//     unsafe fn read<T, const BLOCK_SIZE: usize>(&self, addr_space: u32, ptr: u32) -> [T;
-// BLOCK_SIZE]     where
-//         T: Copy + Debug,
-//     {
-//         debug_assert_eq!(
-//             size_of::<T>(),
-//             self.cell_size[addr_space as usize]
-//         );
-//         let read = self
-//             .paged_vecs
-//             .get_unchecked(addr_space as usize)
-//             .get((ptr as usize) * size_of::<T>());
-//         read
-//     }
-
-//     unsafe fn write<T, const BLOCK_SIZE: usize>(
-//         &mut self,
-//         addr_space: u32,
-//         ptr: u32,
-//         values: &[T; BLOCK_SIZE],
-//     ) where
-//         T: Copy + Debug,
-//     {
-//         debug_assert_eq!(
-//             size_of::<T>(),
-//             self.cell_size[addr_space as usize],
-//             "addr_space={addr_space}"
-//         );
-//         self.paged_vecs
-//             .get_unchecked_mut(addr_space as usize)
-//             .set((ptr as usize) * size_of::<T>(), values);
-//     }
-// }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_basic_get_set() {
-        let mut v = PagedVec::new(3);
-        assert_eq!(v.get::<u32>(0), 0u32);
-        v.set(0, &42u32);
-        assert_eq!(v.get::<u32>(0), 42u32);
-    }
-
-    // TEMP: disable tests (need to update indexing * size_of<u32>)
-    // #[test]
-    // fn test_cross_page_operations() {
-    //     let mut v = PagedVec::<16>::new(3);
-    //     v.set(3, 10u32); // Last element of first page
-    //     v.set(4, 20u32); // First element of second page
-    //     assert_eq!(v.get(3), 10u32);
-    //     assert_eq!(v.get(4), 20u32);
-    // }
-
-    // #[test]
-    // fn test_page_boundaries() {
-    //     let mut v = PagedVec::<16>::new(2);
-    //     // Fill first page
-    //     v.set(0, 1u32);
-    //     v.set(1, 2u32);
-    //     v.set(2, 3u32);
-    //     v.set(3, 4u32);
-    //     // Fill second page
-    //     v.set(4, 5u32);
-    //     v.set(5, 6u32);
-    //     v.set(6, 7u32);
-    //     v.set(7, 8u32);
-
-    //     // Verify all values
-    //     assert_eq!(v.get::<[u32; 8]>(0), [1, 2, 3, 4, 5, 6, 7, 8]);
-    // }
-
-    // #[test]
-    // fn test_range_cross_page_boundary() {
-    //     let mut v = PagedVec::<16>::new(2);
-    //     v.set::<[u32; 6]>(2, [10, 11, 12, 13, 14, 15]);
-    //     assert_eq!(v.get::<[u32; 6]>(2), [10, 11, 12, 13, 14, 15]);
-    // }
-
-    // #[test]
-    // fn test_large_indices() {
-    //     let mut v = PagedVec::<16>::new(100);
-    //     let large_index = 399;
-    //     v.set(large_index, 42u32);
-    //     assert_eq!(v.get(large_index), 42u32);
-    // }
-
-    // #[test]
-    // fn test_range_operations_with_defaults() {
-    //     let mut v = PagedVec::<16>::new(3);
-    //     v.set(2, 5u32);
-    //     v.set(5, 10u32);
-
-    //     // Should include both set values and defaults
-    //     assert_eq!(v.range_vec(1..7), [0, 5, 0, 0, 10, 0]);
-    // }
-
-    // #[test]
-    // fn test_non_zero_default_type() {
-    //     let mut v: PagedVec<4> = PagedVec::new(2);
-    //     assert_eq!(v.get(0), false); // bool's default
-    //     v.set(0, true);
-    //     assert_eq!(v.get(0), true);
-    //     assert_eq!(v.get(1), false); // because we created the page
-    // }
-
-    // #[test]
-    // fn test_set_range_overlapping_pages() {
-    //     let mut v = PagedVec::<_, 16>::new(3);
-    //     let test_data = [1u32, 2, 3, 4, 5, 6];
-    //     v.set(2, test_data);
-
-    //     // Verify first page
-    //     assert_eq!(v.get(2), 1u32);
-    //     assert_eq!(v.get(3), 2u32);
-
-    //     // Verify second page
-    //     assert_eq!(v.get(4), 3u32);
-    //     assert_eq!(v.get(5), 4u32);
-    //     assert_eq!(v.get(6), 5u32);
-    //     assert_eq!(v.get(7), 6u32);
-    // }
-
-    // #[test]
-    // fn test_overlapping_set_ranges() {
-    //     let mut v = PagedVec::<_, 16>::new(3);
-
-    //     // Initial set_range
-    //     v.set(0, [1u32, 2, 3, 4, 5]);
-    //     assert_eq!(v.get::<[u32; 5]>(0), [1, 2, 3, 4, 5]);
-
-    //     // Overlap from beginning
-    //     v.set(0, [10u32, 20, 30]);
-    //     assert_eq!(v.get::<[u32; 5]>(0), [10, 20, 30, 4, 5]);
-
-    //     // Overlap in middle
-    //     v.set(2, [42u32, 43]);
-    //     assert_eq!(v.get::<[u32; 5]>(0), [10, 20, 42, 43, 5]);
-
-    //     // Overlap at end
-    //     v.set(4, [91u32, 92]);
-    //     assert_eq!(v.get::<[u32; 6]>(0), [10, 20, 42, 43, 91, 92]);
-    // }
-
-    // #[test]
-    // fn test_overlapping_set_ranges_cross_pages() {
-    //     let mut v = PagedVec::<16>::new(3);
-
-    //     // Fill across first two pages
-    //     v.set::<[u32; 8]>(0, [1, 2, 3, 4, 5, 6, 7, 8]);
-
-    //     // Overlap end of first page and start of second
-    //     v.set::<[u32; 4]>(2, [21, 22, 23, 24]);
-    //     assert_eq!(v.get::<[u32; 8]>(0), [1, 2, 21, 22, 23, 24, 7, 8]);
-
-    //     // Overlap multiple pages
-    //     v.set::<[u32; 6]>(1, [31, 32, 33, 34, 35, 36]);
-    //     assert_eq!(v.get::<[u32; 8]>(0), [1, 31, 32, 33, 34, 35, 36, 8]);
-    // }
-
-    // #[test]
-    // fn test_iterator() {
-    //     let mut v = PagedVec::<16>::new(3);
-
-    //     v.set(4..10, &[1, 2, 3, 4, 5, 6]);
-    //     let contents: Vec<_> = v.iter().collect();
-    //     assert_eq!(contents.len(), 8); // two pages
-
-    //     contents
-    //         .iter()
-    //         .take(6)
-    //         .enumerate()
-    //         .for_each(|(i, &(idx, val))| {
-    //             assert_eq!((idx, val), (4 + i, 1 + i));
-    //         });
-    //     assert_eq!(contents[6], (10, 0));
-    //     assert_eq!(contents[7], (11, 0));
-    // }
 }
